🤗 Combine SearchSpec and SearchCommandSpec
